pyproject.toml
[project]
name = "spotify-exporter"
version = "0.1.0"
description = "Spotify API authentication using PKCE with spotipy library"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dictor>=0.1.12",
    "fastapi>=0.116.1",
    "pkce>=1.0.3",
    "pytest>=8.4.1",
    "python-dotenv>=1.0.0",
    "requests>=2.32.4",
    "ruff>=0.12.4",
    "spotipy>=2.25.1",
]

[project.scripts]
spotify-exporter = "main:main"

[project.urls]
Homepage = "https://github.com/yourusername/spotify-exporter"
Repository = "https://github.com/yourusername/spotify-exporter"
Issues = "https://github.com/yourusername/spotify-exporter/issues"

[tool.pytest.ini_options]
testpaths = ["test"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py313"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
